beat-ouedkniss/requirements.txt
beautifulsoup4==4.9.0
lxml==4.5.0
Pillow==7.1.2
pytesseract==0.3.4
requests==2.23.0
